requirements.txt
flask
flask-bcrypt
flask-sqlalchemy
psycopg2-binary
python-dotenv